Some rock obstacle attempt (unfinished)
I tried to make rock obstacle with babylonjs's native mesh but they look like shit even with a height map. So I went to import some already made rock-looking .obj meshs.
But it's unfinished, there's a console error and the models are not used yet

Also there's a credit section in README